homework/fastapi-demo/backend/requirements.txt
fastapi[standard]
mysql-connector-python
pydantic
